static project
In the app2.py file, we pull the data in the .csv file and display the project in the form of 2 static world maps. The first of these maps shows the speed value as a Rainbow on the world map, and the other shows the source country and the current location of the aircraft as a route. We have completed our project.